[nvim] migrate from packer to lazy.nvim
