+++ .chezmoiremove
@@ -2,3 +2,4 @@
 .config/electron19-flags.conf
 .config/electron24-flags.conf
 .config/electron25-flags.conf
+.config/nvim/after/
