--- dot_config/nvim/after/plugin/neosolarized.lua
@@ -1,3 +1,0 @@
-require('neosolarized').setup({
-  background_set = false,
-})
