--- dot_config/nvim/after/plugin/neotest.lua
@@ -1,9 +1,0 @@
-require("neotest").setup({
-  adapters = {
-    require("neotest-python")({ 
-      -- shitcode warning
-      python = vim.fn.trim(vim.fn.system('poetry env info -p 2> /dev/null')) .. '/bin/python'
-    })
-  }
-})
-
