--- dot_config/nvim/after/plugin/telescope.lua
@@ -1,2 +1,0 @@
-require('telescope').load_extension('fzf')
-
