+++ dot_config/nvim/lua/n0nvme/init.lua
@@ -1,0 +1,21 @@
+require("n0nvme.keys")
+require("n0nvme.options")
+
+local lazypath = vim.fn.stdpath("data") .. "/lazy/lazy.nvim"
+if not vim.loop.fs_stat(lazypath) then
+  vim.fn.system({
+    "git",
+    "clone",
+    "--filter=blob:none",
+    "https://github.com/folke/lazy.nvim.git",
+    "--branch=stable", -- latest stable release
+    lazypath,
+  })
+end
+vim.opt.rtp:prepend(lazypath)
+
+require("lazy").setup({
+    spec = "n0nvme.plugins",
+    change_detection = { notify = false }
+})
+
