+++ dot_config/nvim/lua/n0nvme/keys.lua
@@ -1,0 +1,2 @@
+vim.g.mapleader = " "
+
