+++ dot_config/nvim/lua/n0nvme/options.lua
@@ -1,0 +1,9 @@
+vim.opt.termguicolors = true
+vim.opt.nu = true
+vim.opt.rnu = true
+vim.opt.syntax = on
+vim.loader.enable()
+
+vim.g.loadednetrw = 1
+vim.g.loaded_netrwPlugin = 1
+
