+++ dot_config/nvim/lua/n0nvme/plugins/telescope.lua
@@ -1,0 +1,17 @@
+return {
+  "nvim-telescope/telescope.nvim",
+  dependencies = {
+	  {"nvim-lua/plenary.nvim"}, 
+	  {
+		  "nvim-telescope/telescope-fzf-native.nvim",
+	            build = 'cmake -S. -Bbuild -DCMAKE_BUILD_TYPE=Release && cmake --build build --config Release && cmake --install build --prefix build'}
+  },
+  config = function()
+	  require("telescope").load_extension("fzf")
+	  local builtin = require('telescope.builtin')
+	  vim.keymap.set('n', '<leader>ff', builtin.find_files, {})
+	  vim.keymap.set('n', '<leader>fg', builtin.live_grep, {})
+	  vim.keymap.set('n', '<leader>fb', builtin.buffers, {})
+	  vim.keymap.set('n', '<leader>fh', builtin.help_tags, {})
+  end
+}
